+++ pydantic_jsonschema/__init__.py
@@ -2,10 +2,8 @@
 
 from pydantic_jsonschema._version import __version__
 from pydantic_jsonschema.converters import (
-    LaxSchemaConverter,
     Ref,
     SchemaConverter,
-    to_lax_model,
     to_model,
 )
 from pydantic_jsonschema.types import DataType, JsonType, Reference, Schema
@@ -13,12 +11,10 @@ from pydantic_jsonschema.types import DataType, JsonType, Reference, Schema
 __all__ = [
     "DataType",
     "JsonType",
-    "LaxSchemaConverter",
     "Ref",
     "Reference",
     "Schema",
     "SchemaConverter",
     "__version__",
-    "to_lax_model",
     "to_model",
 ]